+++ scripts/legacyMacros.js
@@ -661,6 +661,7 @@ export {charlatansDie} from './macros/2014/items/trinket/charlatansDie.js';
 export {deckOfMiscellany, deckOfMiscellanySetup, miscellanyCard} from './macros/2014/items/trinket/deckOfMiscellany.js';
 export {deckOfWonder, deckOfWonderDestinyEffect, deckOfWonderIsolationEffect, deckOfWonderJusticeEffect, deckOfWonderVultureLostEffect} from './macros/2014/items/trinket/deckOfWonder.js';
 export {harkonsBite, harkonsBiteEffect} from './macros/2014/items/trinket/harkonsBite.js';
+export {crownOfDaffodils} from './macros/2014/items/trinket/crownOfDaffodils.js';
 //BG3 Weapon Actions
 export {backbreaker, braceMelee, braceMeleeDamage, cleave, concussiveSmash, maimingStrike, disarmingStrike, flourish, heartstopper, lacerate, piercingStrike, piercingStrikeEffect, pommelStrike, prepare, rush, tenacity, topple, weakeningStrike, braceRanged, braceRangedDamage, hamstringShot, mobileShot, piercingShot} from './macros/2014/homebrew/bg3WeaponActions.js';
 //Other