+++ scripts/macros/2014/spells/conjureAnimals.js
@@ -25,12 +25,12 @@ async function use({workflow}) {
         return;
     }
     let userId = socketUtils.gmID();
-    if (genericUtils.getCPRSetting('playerSelectsConjures')) userId = game.userId;
+    if (genericUtils.getCPRSetting('playerSelectsConjures')) userId = game.user.id;
     if (!userId) {
         if (concentrationEffect) await genericUtils.remove(concentrationEffect);
         return;
     }
-    let sourceDocs = await dialogUtils.selectDocumentsDialog(workflow.item.name, genericUtils.format('CHRISPREMADES.Summons.SelectSummons', {totalSummons: totalSummons * 2 / selection}), compendiumDocs, {max: totalSummons * 2 / selection, sortAlphabetical: true, sortCR: true, showCR: true});
+    let sourceDocs = await dialogUtils.selectDocumentsDialog(workflow.item.name, genericUtils.format('CHRISPREMADES.Summons.SelectSummons', {totalSummons: totalSummons * 2 / selection}), compendiumDocs, {max: totalSummons * 2 / selection, sortAlphabetical: true, sortCR: true, showCR: true, userId});
     if (sourceDocs?.length) sourceDocs = sourceDocs?.filter(i => i.amount);
     if (!sourceDocs?.length) {
         if (concentrationEffect) await genericUtils.remove(concentrationEffect);