+++ scripts/macros/2024/monsterFeatures/vampire/graveStrike.js
@@ -1,0 +1,7 @@
+export let graveStrike = {
+    name: 'Grave Strike (Vampire Form Only)',
+    aliases: ['Grave Strike'],
+    monster: 'Vampire',
+    version: '1.3.126',
+    rules: 'modern'
+};